+++ test/data/helmholtz/mesh/test_crystal_builders.py
@@ -1,6 +1,35 @@
-import gmsh
+import pathlib
 
+import gmsh
+import pytest
+
+from nos.data.helmholtz.domain_properties import CShapeDescription, CylinderDescription, NoneDescription, read_config
 from nos.data.helmholtz.mesh import CrystalBuilder, CShapedCrystalBuilder, CylindricalCrystalBuilder
+
+
+@pytest.fixture
+def descriptions():
+    templates = pathlib.Path.cwd().joinpath("templates")
+    files = templates.glob("*.ini")
+    descriptions = []
+    for file in files:
+        descriptions.extend(read_config(file))
+    return descriptions
+
+
+@pytest.fixture
+def cylindrical_descriptions(descriptions):
+    return [des for des in descriptions if isinstance(des.crystal_description, CylinderDescription)]
+
+
+@pytest.fixture
+def c_shaped_descriptions(descriptions):
+    return [des for des in descriptions if isinstance(des.crystal_description, CShapeDescription)]
+
+
+@pytest.fixture
+def none_descriptions(descriptions):
+    return [des for des in descriptions if isinstance(des.crystal_description, NoneDescription)]
 
 
 def test_base_crystal_builder_empty(descriptions):
